6646: move the introduced view model classes to a separate idl and implementation + add documentation + handle the case where we cannot cast propeprty args
