+++ src/cascadia/TerminalApp/CommandPalette.idl
@@ -4,20 +4,10 @@
 import "IDirectKeyListener.idl";
 import "ShortcutActionDispatch.idl";
 import "HighlightedTextControl.idl";
+import "FilteredCommand.idl";
 
 namespace TerminalApp
 {
-    [default_interface] runtimeclass FilteredCommand : Windows.UI.Xaml.Data.INotifyPropertyChanged
-    {
-        FilteredCommand();
-        FilteredCommand(Microsoft.Terminal.Settings.Model.Command command);
-
-        Microsoft.Terminal.Settings.Model.Command Command { get; };
-        String Filter;
-        HighlightedText HighlightedName { get; };
-        void UpdateFilter(String filter);
-    }
-
     [default_interface] runtimeclass CommandPalette : Windows.UI.Xaml.Controls.UserControl, Windows.UI.Xaml.Data.INotifyPropertyChanged, IDirectKeyListener
     {
         CommandPalette();
