+++ src/cascadia/TerminalApp/FilteredCommand.idl
@@ -1,0 +1,20 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import "IDirectKeyListener.idl";
+import "ShortcutActionDispatch.idl";
+import "HighlightedTextControl.idl";
+
+namespace TerminalApp
+{
+    [default_interface] runtimeclass FilteredCommand : Windows.UI.Xaml.Data.INotifyPropertyChanged
+    {
+        FilteredCommand();
+        FilteredCommand(Microsoft.Terminal.Settings.Model.Command command);
+
+        Microsoft.Terminal.Settings.Model.Command Command { get; };
+        String Filter;
+        HighlightedText HighlightedName { get; };
+        void UpdateFilter(String filter);
+    }
+}
