+++ src/cascadia/TerminalApp/TerminalAppLib.vcxproj.filters
@@ -27,6 +27,12 @@
     <ClCompile Include="TerminalSettings.cpp">
       <Filter>settings</Filter>
     </ClCompile>
+    <ClCompile Include="FilteredCommand.cpp">
+      <Filter>commandPalette</Filter>
+    </ClCompile>
+    <ClCompile Include="HighlightedText.cpp">
+      <Filter>highlightedText</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Utils.h" />
@@ -47,6 +53,12 @@
     <ClInclude Include="ColorHelper.h" />
     <ClInclude Include="TerminalSettings.h">
       <Filter>settings</Filter>
+    </ClInclude>
+    <ClInclude Include="FilteredCommand.h">
+      <Filter>commandPalette</Filter>
+    </ClInclude>
+    <ClInclude Include="HighlightedText.h">
+      <Filter>highlightedText</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -69,6 +81,12 @@
     <Midl Include="CommandKeyChordVisibilityConverter.idl" />
     <Midl Include="TerminalSettings.idl">
       <Filter>settings</Filter>
+    </Midl>
+    <Midl Include="FilteredCommand.idl">
+      <Filter>commandPalette</Filter>
+    </Midl>
+    <Midl Include="HighlightedText.idl">
+      <Filter>highlightedText</Filter>
     </Midl>
   </ItemGroup>
   <ItemGroup>
@@ -94,7 +112,7 @@
       <Filter>commandPalette</Filter>
     </Page>
     <Page Include="HighlightedTextControl.xaml">
-      <Filter>controls</Filter>
+      <Filter>highlightedText</Filter>
     </Page>
   </ItemGroup>
   <ItemGroup>
@@ -115,6 +133,9 @@
     </Filter>
     <Filter Include="commandPalette">
       <UniqueIdentifier>{2ad498e1-d8ea-4381-9464-a74c141bd7dd}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="highlightedText">
+      <UniqueIdentifier>{e490f626-547d-4b5b-b22d-c6d33c9e3210}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
